+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Summary of release changes for Version 2 - CentOS-7
 ### 2.3.2 - Unreleased
 
 - Updates supervisor to 3.3.4.
+- Adds feature to set `SSH_USER_PASSWORD` via a file path. i.e Docker secrets.
 
 ### 2.3.1 - 2018-01-12